--- src/contexts/hotReloadContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const HotReloadContext = createContext(false);
